cambiados dos valores float a double por consistencia

+++ app/src/main/java/com/example/intermodular/core/route/client/single/RouteResponse.kt
@@ -17,8 +17,8 @@ data class RouteResponse (
 )
 
 data class Location (
-    @SerializedName("latitude")         val latitude: Float,
-    @SerializedName("longitude")        val longitude: Float,
+    @SerializedName("latitude")         val latitude: Double,
+    @SerializedName("longitude")        val longitude: Double,
     @SerializedName("waypoint")         val waypoint: Waypoint?
 )
